backend/turantalim-backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-extensions==3.2.3
django-jazzmin==3.0.1
# psycopg2-binary==2.9.9  # PostgreSQL driver - uncomment if using PostgreSQL
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.29.7
requests==2.31.0
PyJWT
cryptography==41.0.7
openai==1.3.5
librosa==0.10.0
soundfile==0.12.1
python-docx==0.8.11
Pillow
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
django-model-utils==4.3.1
django-timezone-field==6.1.0
python-telegram-bot==20.3
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
